RavenDB-20504: Applied review request.
