+++ src/Raven.Server/Documents/Indexes/Index.cs
@@ -120,7 +120,7 @@ namespace Raven.Server.Documents.Indexes
 
         private readonly Size MappedSizeLimitOn32Bits = new Size(8, SizeUnit.Megabytes);
 
-        private const int IndexingCompressionMaxTestDocuments = 10000;
+
 
         protected Logger _logger;
 
@@ -162,7 +162,7 @@ namespace Raven.Server.Documents.Indexes
 
         private bool _initialized;
 
-        protected UnmanagedBuffersPoolWithLowMemoryHandling _unmanagedBuffersPool;
+        internal UnmanagedBuffersPoolWithLowMemoryHandling _unmanagedBuffersPool;
 
         internal StorageEnvironment _environment;
 
@@ -889,139 +889,6 @@ namespace Raven.Server.Documents.Indexes
             }
 
             IndexPersistence.Initialize(environment);
-            EnsureIsInitialized();
-        }
-
-        private void EnsureIsInitialized()
-        {
-            if (SearchEngineType == SearchEngineType.Lucene)
-                return;
-
-            using (_contextPool.AllocateOperationContext(out TransactionOperationContext context))
-            using (var tx = context.OpenReadTransaction())
-            {
-                // It is already initialized, we can return.
-                if (CompactTree.HasDictionary(tx.InnerTransaction.LowLevelTransaction))
-                    return;
-            }
-
-            // We know that Corax is not initialized yet, therefore we will call 
-            OnCoraxInitialization();
-        }
-
-        private struct DocumentTrainSpanEnumerator : IReadOnlySpanEnumerator
-        {
-            private readonly DocumentsStorage _documentStorage;
-            private readonly QueryOperationContext _queryContext;
-            private readonly TransactionOperationContext _indexContext;
-            private readonly Index _index;
-            private readonly IndexType _indexType;
-            private readonly HashSet<string> _collections;
-            private readonly int _take;
-            private IEnumerator<LazyStringValue> _itemsEnumerable;
-
-            public DocumentTrainSpanEnumerator(TransactionOperationContext indexContext, Index index, IndexType indexType, DocumentsStorage storage, QueryOperationContext queryContext, HashSet<string> collections, int take = int.MaxValue)
-            {
-                _indexContext = indexContext;
-                _index = index;
-                _indexType = indexType;
-                _take = take;
-
-                _documentStorage = storage;
-                _queryContext = queryContext;
-                _collections = collections;
-            }
-
-            private IEnumerable<LazyStringValue> GetItems()
-            {
-                var scope = new IndexingStatsScope(new IndexingRunStats());
-                foreach (var collection in _collections)
-                {
-                    using var itemEnumerator = _index.GetMapEnumerator(GetItemsEnumerator(_queryContext, collection, _take), collection, _indexContext, scope, _indexType);
-                    while (true)
-                    {
-                        if (itemEnumerator.MoveNext(_queryContext.Documents, out var _, out var _) == false)
-                            break;
-
-                        Document doc = itemEnumerator.Current.Item as Document;
-                        if (doc == null)
-                            continue;
-
-                        var reader = doc.Data;
-                        int properties = reader.Count;
-                        BlittableJsonReaderObject.PropertyDetails details = default;
-                        for (int i = 0; i < properties; i++)
-                        {
-                            reader.GetPropertyByIndex(i, ref details);
-                            switch (details.Token)
-                            {
-                                case BlittableJsonToken.String:
-                                    yield return (LazyStringValue)details.Value;
-                                    break;
-                                case BlittableJsonToken.CompressedString:
-                                    yield return ((LazyCompressedStringValue)details.Value).ToLazyStringValue();
-                                    break;
-                            }
-                        }
-                    }
-                }
-            }
-
-            private IEnumerable<Document> GetDocumentsEnumerator(QueryOperationContext queryContext, string collection, long take = int.MaxValue)
-            {
-                if (collection == Constants.Documents.Collections.AllDocumentsCollection)
-                    return _documentStorage.GetUniformlyDistributedDocumentsFrom(queryContext.Documents, take);
-                return _documentStorage.GetUniformlyDistributedDocumentsFrom(queryContext.Documents, collection, take);
-            }
-
-            private IEnumerable<IndexItem> GetItemsEnumerator(QueryOperationContext queryContext, string collection, long take = int.MaxValue)
-            {
-                foreach (var document in GetDocumentsEnumerator(queryContext, collection, take))
-                {
-                    yield return new DocumentIndexItem(document.Id, document.LowerId, document.Etag, document.LastModified, document.Data.Size, document);
-                }
-            }
-
-            public void Reset()
-            {
-                _itemsEnumerable = GetItems().GetEnumerator();
-            }
-
-            public bool MoveNext(out ReadOnlySpan<byte> output)
-            {
-                _itemsEnumerable ??= GetItems().GetEnumerator();
-                var result = _itemsEnumerable.MoveNext();
-                output = result == true? _itemsEnumerable.Current.AsSpan() : ReadOnlySpan<byte>.Empty;
-                return result;
-            }
-        }
-
-        protected virtual void OnCoraxInitialization()
-        {
-            var documentStorage = DocumentDatabase.DocumentsStorage;
-
-            using (CultureHelper.EnsureInvariantCulture())
-            using (_contextPool.AllocateOperationContext(out TransactionOperationContext indexContext))
-            using (var queryContext = QueryOperationContext.Allocate(DocumentDatabase, this))
-            using (CurrentIndexingScope.Current = new CurrentIndexingScope(this, documentStorage, queryContext, Definition, indexContext, GetOrAddSpatialField, _unmanagedBuffersPool))
-            {
-                indexContext.PersistentContext.LongLivedTransactions = true;
-                queryContext.SetLongLivedTransactions(true);
-                queryContext.OpenReadTransaction();
-
-                var tx = indexContext.OpenWriteTransaction();
-
-                var enumerator = new DocumentTrainSpanEnumerator(indexContext, this, Type, documentStorage, queryContext, Collections, Configuration.CoraxIndexingMaxDocumentsForDictionary);
-                var testEnumerator = new DocumentTrainSpanEnumerator(indexContext, this, Type, documentStorage, queryContext, Collections, IndexingCompressionMaxTestDocuments);
-
-                var llt = tx.InnerTransaction.LowLevelTransaction;
-                var defaultDictionaryId = PersistentDictionary.CreateDefault(llt);
-                var defaultDictionary = new PersistentDictionary(llt.GetPage(defaultDictionaryId));
-
-                PersistentDictionary.ReplaceIfBetter(llt, enumerator, testEnumerator, defaultDictionary);
-
-                tx.Commit();
-            }
         }
 
         protected virtual void OnInitialization()
@@ -5201,7 +5068,7 @@ namespace Raven.Server.Documents.Indexes
             return _regexCache.Get(arg);
         }
 
-        protected SpatialField GetOrAddSpatialField(string name)
+        internal SpatialField GetOrAddSpatialField(string name)
         {
             return _spatialFields.GetOrAdd(name, n =>
             {
